# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for app.",
        "title": "ConformeTec",
        "contact": {},
        "version": "1.0"
    },
    "host": "3.138.100.192:8080",
    "paths": {
        "/all-find-painel": {
            "get": {
                "description": "Procurar um Painel existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Puxa todos os painels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.SectorTest"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/api/test-token": {
            "post": {
                "description": "Cria um toke para auth do usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Test Token",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Token de autenticação",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token make:",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/create-painels-test": {
            "post": {
                "description": "Criar um novo painel para uma empresa",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Criar painel Test",
                "parameters": [
                    {
                        "description": "Dados do campo a ser criado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.SectorInputTest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.SectorTest"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/create-sector": {
            "post": {
                "description": "Criar um novo setor para uma empresa",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sector"
                ],
                "summary": "Criar setor",
                "parameters": [
                    {
                        "description": "Dados do campo a ser criado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.SectorInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Sector"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/find-company": {
            "get": {
                "description": "Procurar uma companhia existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Procurar uma companhia",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company do usuario",
                        "name": "Company",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Company"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/find-sector": {
            "get": {
                "description": "Procurar um setor especifico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sector"
                ],
                "summary": "Procurar setor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company do usuario",
                        "name": "Company",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nome do setor",
                        "name": "NameSector",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Numero do setor",
                        "name": "NumberSector",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Sector"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/find-user": {
            "get": {
                "description": "Procurar um usuario existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Procurar Usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rg do usuario",
                        "name": "RG",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.User"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/insert-nc": {
            "post": {
                "description": "Insere um novo registro NC",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NC"
                ],
                "summary": "Inserir NC",
                "parameters": [
                    {
                        "description": "Dados de entrada para inserção",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.NotConformInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File uploaded successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/loggout": {
            "put": {
                "description": "Altera o modo de login para false",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Desconectar o usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email do usuario",
                        "name": "Email",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/login": {
            "put": {
                "description": "Altera o modo de login para true",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Conectar o usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email do usuario",
                        "name": "Email",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Senha do usuario",
                        "name": "Password",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/pull-nc": {
            "get": {
                "description": "Obtém registros NC com base nos parâmetros fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NC"
                ],
                "summary": "Obter NC",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email do usuario",
                        "name": "Email",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Criar um usuario a partir das novas infos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Criar um novo usuario",
                "parameters": [
                    {
                        "description": "Dados do campo a ser criado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.UserInput"
                        }
                    },
                    {
                        "enum": [
                            10,
                            11,
                            12,
                            13
                        ],
                        "type": "integer",
                        "description": "Tipo de usuário (Gerente da Planta 10, Profissional habilitado responsável 11, Responsável pelo PIE 12, Profissional autorizado 13)",
                        "name": "Office",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.User"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/register-company": {
            "post": {
                "description": "Criar uma nova companhia",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Criar uma nova companhia",
                "parameters": [
                    {
                        "description": "Dados do campo a ser criado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.CompanyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Company"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
        "/token": {
            "get": {
                "description": "Cria um toke para auth do usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Criar token de auth",
                "responses": {
                    "200": {
                        "description": "token make:",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "base.BaseReq": {
            "type": "object",
            "properties": {
                "LogData": {
                    "$ref": "#/definitions/base.LogData"
                },
                "ResponseData": {
                    "$ref": "#/definitions/base.ResponseData"
                }
            }
        },
        "base.LogData": {
            "type": "object",
            "properties": {
                "ExecutionTime": {
                    "type": "string"
                },
                "HttpStatusCode": {
                    "type": "integer"
                },
                "IP": {
                    "type": "string"
                },
                "Route": {
                    "type": "string"
                }
            }
        },
        "base.ResponseData": {
            "type": "object",
            "properties": {
                "Response": {}
            }
        },
        "db.Company": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "string"
                },
                "cnpj": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "db.Sector": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                }
            }
        },
        "db.SectorTest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                }
            }
        },
        "db.User": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isLogged": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "erros.InternalServerError": {
            "type": "object",
            "properties": {
                "BaseReq": {
                    "$ref": "#/definitions/base.BaseReq"
                },
                "InternalServerError": {
                    "$ref": "#/definitions/erros.Message"
                },
                "Response": {}
            }
        },
        "erros.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "interfaces.CompanyInput": {
            "type": "object",
            "required": [
                "area",
                "cnpj",
                "company",
                "email",
                "name",
                "password",
                "type"
            ],
            "properties": {
                "area": {
                    "type": "string"
                },
                "cnpj": {
                    "type": "string"
                },
                "company": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "interfaces.NotConformInput": {
            "type": "object",
            "properties": {
                "Desc": {
                    "type": "string"
                },
                "Legal": {
                    "type": "string"
                },
                "Number": {
                    "type": "integer"
                },
                "Recomended": {
                    "type": "string"
                },
                "Tech": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                }
            }
        },
        "interfaces.SectorInput": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                }
            }
        },
        "interfaces.SectorInputTest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                }
            }
        },
        "interfaces.UserInput": {
            "type": "object",
            "required": [
                "companyName",
                "email",
                "name",
                "office",
                "password"
            ],
            "properties": {
                "companyName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "office": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}